remove artifact functionality
